% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{add_centroid_coordinates}
\alias{add_centroid_coordinates}
\title{Add Centroid Coordinates to Spatial Data}
\usage{
add_centroid_coordinates(spatial_data)
}
\arguments{
\item{spatial_data}{sf object. Spatial data}
}
\value{
Data frame with centroid coordinates added
}
\description{
Calculate centroid coordinates for spatial units
}
