% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{check_builtin_data}
\alias{check_builtin_data}
\title{Check Data Availability from OSF}
\usage{
check_builtin_data(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Show detailed information about cache status}
}
\value{
List showing available datasets and cache status
}
\description{
Check Data Availability from OSF
}
