% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text Data}
\usage{
clean_text(text)
}
\arguments{
\item{text}{Character vector to clean}
}
\value{
Character vector of cleaned text
}
\description{
Remove extra quotes and whitespace from text fields
}
