% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{clear_data_cache}
\alias{clear_data_cache}
\title{Clear Data Cache}
\usage{
clear_data_cache(confirm = TRUE, verbose = TRUE)
}
\arguments{
\item{confirm}{Logical. Require confirmation before deleting}

\item{verbose}{Logical. Show what's being deleted}
}
\value{
Logical. TRUE if successful
}
\description{
Remove cached datasets to free up disk space
}
