% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{create_batch_summary_report}
\alias{create_batch_summary_report}
\title{Create Batch Summary Report}
\usage{
create_batch_summary_report(
  batch_results,
  years,
  nutrients,
  output_dir,
  verbose
)
}
\arguments{
\item{batch_results}{List of results}

\item{years}{Vector of years}

\item{nutrients}{Vector of nutrients}

\item{output_dir}{Output directory}

\item{verbose}{Show progress}
}
\value{
Summary data frame
}
\description{
Generate comprehensive summary report for batch analysis
}
\keyword{internal}
