% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{create_classification_summary}
\alias{create_classification_summary}
\title{Create Classification Summary Table}
\usage{
create_classification_summary(data, agricultural_col, combined_col)
}
\arguments{
\item{data}{Data frame. Data with classification columns}

\item{agricultural_col}{Character. Name of agricultural classification column}

\item{combined_col}{Character. Name of combined (with WWTP) classification column}
}
\value{
Data frame with classification counts and changes
}
\description{
Create summary table of classification counts for both nutrients
}
