% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{create_classification_trend_plot}
\alias{create_classification_trend_plot}
\title{Create Classification Trend Plot}
\usage{
create_classification_trend_plot(trend_data, nutrient)
}
\arguments{
\item{trend_data}{Data frame with trend data}

\item{nutrient}{Character nutrient type}
}
\value{
ggplot object
}
\description{
Create Classification Trend Plot
}
\keyword{internal}
