% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{create_comparison_plots}
\alias{create_comparison_plots}
\title{Create Comparison Plots}
\usage{
create_comparison_plots(comparison_data, metrics)
}
\arguments{
\item{comparison_data}{Data frame with scenario metrics}

\item{metrics}{Character vector of metrics to plot}
}
\value{
List of ggplot objects
}
\description{
Create Comparison Plots
}
\keyword{internal}
