% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_comparison_functions.R
\name{create_regional_plots}
\alias{create_regional_plots}
\title{Create Regional Comparison Plots}
\usage{
create_regional_plots(summary_data, nutrients)
}
\description{
Create Regional Comparison Plots
}
\keyword{internal}
