% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_comparison_functions.R
\name{create_regional_summary}
\alias{create_regional_summary}
\title{Create Regional Summary}
\usage{
create_regional_summary(regional_results, nutrients)
}
\description{
Create Regional Summary
}
\keyword{internal}
