% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{create_stacked_area_plot}
\alias{create_stacked_area_plot}
\title{Create Stacked Area Plot}
\usage{
create_stacked_area_plot(trend_data, nutrient)
}
\arguments{
\item{trend_data}{Data frame with trend data}

\item{nutrient}{Character nutrient type}
}
\value{
ggplot object
}
\description{
Create Stacked Area Plot
}
\keyword{internal}
