% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{create_yoy_change_plot}
\alias{create_yoy_change_plot}
\title{Create Year-over-Year Change Plot}
\usage{
create_yoy_change_plot(trend_data, nutrient)
}
\arguments{
\item{trend_data}{Data frame with trend data}

\item{nutrient}{Character nutrient type}
}
\value{
ggplot object
}
\description{
Create Year-over-Year Change Plot
}
\keyword{internal}
