% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Startup Message for OSF Data Approach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Character. Library name (passed by R)}

\item{pkgname}{Character. Package name (passed by R)}
}
\description{
Displays informative startup message when package is loaded
}
\keyword{internal}
