% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\name{.package_info}
\alias{.package_info}
\title{Package Version and Build Information}
\usage{
.package_info()
}
\value{
List with package metadata
}
\description{
Internal function to provide package build information
}
\keyword{internal}
