% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{download_all_data}
\alias{download_all_data}
\title{Download All Datasets from OSF}
\usage{
download_all_data(force_download = FALSE, verbose = TRUE)
}
\arguments{
\item{force_download}{Logical. Re-download even if files exist in cache}

\item{verbose}{Logical. Show progress for each download}
}
\value{
Logical. TRUE if all downloads successful
}
\description{
Convenience function to download all available datasets from OSF
}
