% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{download_osf_data}
\alias{download_osf_data}
\title{Download and Cache Data from OSF}
\usage{
download_osf_data(
  dataset_name,
  force_download = FALSE,
  cache_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dataset_name}{Character. Name of dataset to download}

\item{force_download}{Logical. Force re-download even if cached version exists}

\item{cache_dir}{Character. Directory to cache downloaded data (default: user data dir)}

\item{verbose}{Logical. Show download progress}
}
\value{
Path to cached data file
}
\description{
Download built-in datasets from OSF repository using Files API
}
