% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{export_for_policy}
\alias{export_for_policy}
\title{Export for Policy Briefs}
\usage{
export_for_policy(results, output_dir = file.path(tempdir(), "policy_export"))
}
\arguments{
\item{results}{Analysis results}

\item{output_dir}{Output directory}
}
\value{
List of created files
}
\description{
Create simplified outputs for policy makers
}
