% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{extract_trend_data}
\alias{extract_trend_data}
\title{Extract Trend Data from Batch Results}
\usage{
extract_trend_data(batch_results, nutrient)
}
\arguments{
\item{batch_results}{List of results}

\item{nutrient}{Character nutrient type}
}
\value{
Data frame with trend data
}
\description{
Extract Trend Data from Batch Results
}
\keyword{internal}
