% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_comparison_functions.R
\name{get_region_definitions}
\alias{get_region_definitions}
\title{Get Built-in Region Definitions}
\usage{
get_region_definitions()
}
\value{
List of region definitions
}
\description{
Get Built-in Region Definitions
}
