% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{get_state_boundaries}
\alias{get_state_boundaries}
\title{Get State Boundaries for Mapping}
\usage{
get_state_boundaries()
}
\value{
sf object with state boundaries
}
\description{
Get US state boundaries excluding non-CONUS states
}
