% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_analysis_functions.R
\name{get_state_fips}
\alias{get_state_fips}
\title{Get State FIPS Code}
\usage{
get_state_fips(state_abbr)
}
\arguments{
\item{state_abbr}{Character. Two-letter state abbreviation (e.g., "OH", "TX")}
}
\value{
Character. Two-digit state FIPS code
}
\description{
Convert state abbreviation to FIPS code
}
