% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_functions.R
\name{load_texas_huc8_boundaries}
\alias{load_texas_huc8_boundaries}
\title{Load Texas HUC8 Boundaries from OSF}
\usage{
load_texas_huc8_boundaries(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Show loading progress}
}
\value{
sf object with Texas HUC8 boundaries
}
\description{
Load Texas HUC8 spatial boundaries from OSF repository
}
\keyword{internal}
