% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_functions.R
\name{load_texas_huc8_data}
\alias{load_texas_huc8_data}
\title{Load Texas HUC8 Data from OSF}
\usage{
load_texas_huc8_data(year = 2016, verbose = FALSE)
}
\arguments{
\item{year}{Numeric. Year to extract from Texas data}

\item{verbose}{Logical. Show loading progress}
}
\value{
Data frame with Texas HUC8 data for specified year
}
\description{
Load Texas HUC8 supplemental data from OSF repository
}
\keyword{internal}
