% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_agricultural_classification}
\alias{map_agricultural_classification}
\title{Create Agricultural Classification Map}
\usage{
map_agricultural_classification(data, nutrient, classification_col, title)
}
\arguments{
\item{data}{sf object. Spatial data with classifications}

\item{nutrient}{Character. "nitrogen" or "phosphorus"}

\item{classification_col}{Character. Name of classification column}

\item{title}{Character. Map title}
}
\value{
ggplot object
}
\description{
Create map showing agricultural nutrient classifications
}
