% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_wwtp_influence}
\alias{map_wwtp_influence}
\title{Create WWTP Influence Map}
\usage{
map_wwtp_influence(data, nutrient, title)
}
\arguments{
\item{data}{sf object. Integrated data with WWTP proportions}

\item{nutrient}{Character. "nitrogen" or "phosphorus"}

\item{title}{Character. Map title}
}
\value{
ggplot object
}
\description{
Create map showing WWTP contribution as proportion of total nutrient load
}
