% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_wwtp_points}
\alias{map_wwtp_points}
\title{Create WWTP Point Map}
\usage{
map_wwtp_points(wwtp_sf, nutrient, title)
}
\arguments{
\item{wwtp_sf}{sf object. Spatial WWTP data with classifications}

\item{nutrient}{Character. "nitrogen" or "phosphorus"}

\item{title}{Character. Map title}
}
\value{
ggplot object
}
\description{
Create map showing WWTP locations classified by load size
}
