% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{plot_absolute_changes}
\alias{plot_absolute_changes}
\title{Create Absolute Change Plot}
\usage{
plot_absolute_changes(data, title)
}
\arguments{
\item{data}{Data frame. Summary data with absolute changes}

\item{title}{Character. Plot title}
}
\value{
ggplot object
}
\description{
Create plot showing absolute changes in classification counts
}
