% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{plot_impact_ratios}
\alias{plot_impact_ratios}
\title{Create Impact Ratio Plot}
\usage{
plot_impact_ratios(data, title)
}
\arguments{
\item{data}{Data frame. Summary data with impact ratios}

\item{title}{Character. Plot title}
}
\value{
ggplot object
}
\description{
Create plot showing impact of WWTP addition as ratios
}
