% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{test_osf_connection}
\alias{test_osf_connection}
\title{Test OSF Connection}
\usage{
test_osf_connection(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Show detailed test results}
}
\value{
Logical. TRUE if test successful
}
\description{
Test downloading a small dataset to verify OSF connectivity
}
