% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate Required Columns}
\usage{
validate_columns(data, required_cols, data_type = "data")
}
\arguments{
\item{data}{Data frame to validate}

\item{required_cols}{Character vector of required column names}

\item{data_type}{Character description of data type for error messages}
}
\value{
Logical. TRUE if all columns present, stops with error otherwise
}
\description{
Validate Required Columns
}
