% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwtp_functions.R
\name{wwtp_aggregate_by_boundaries}
\alias{wwtp_aggregate_by_boundaries}
\title{Aggregate WWTP Data by Spatial Boundaries}
\usage{
wwtp_aggregate_by_boundaries(wwtp_sf, boundaries, nutrient, boundary_id_col)
}
\arguments{
\item{wwtp_sf}{sf object. Spatial WWTP data}

\item{boundaries}{sf object. Spatial boundaries for aggregation}

\item{nutrient}{Character. "nitrogen" or "phosphorus"}

\item{boundary_id_col}{Character. Name of boundary ID column}
}
\value{
Data frame with aggregated loads by spatial unit
}
\description{
Aggregate WWTP loads by spatial units (counties, HUC8s, etc.)
}
