% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwtp_functions.R
\name{wwtp_filter_positive_loads}
\alias{wwtp_filter_positive_loads}
\title{Filter WWTP Data for Positive Loads}
\usage{
wwtp_filter_positive_loads(wwtp_data, nutrient)
}
\arguments{
\item{wwtp_data}{Data frame. WWTP data}

\item{nutrient}{Character. "nitrogen" or "phosphorus"}
}
\value{
Data frame with facilities having positive loads
}
\description{
Filter WWTP data to include only facilities with positive nutrient loads
}
